requirements.txt
aws-cdk-lib==2.142.1
constructs>=10.0.0,<11.0.0
aws_cdk.aws_lambda_python_alpha
aws-cdk.aws-lambda-url